split search filter utils into new file
